[refactor]: 여행지 삭제 로직 수정 (#72) - 삭제 훅을 유틸 함수로 변경 - 예외처리를 마이페이지 컴포넌트에서 진행
