+++ src/app/myJourney/page.tsx
@@ -1,29 +1,38 @@
 'use client';
 
-import { useContext } from 'react';
+import { useContext, useState } from 'react';
 import { Button, Spinner } from '@nextui-org/react';
 import { useRouter } from 'next/navigation';
 
 import MyPlaceCard from '@/components/Card/MyPlaceCard';
 import PaginationBar from '@/components/Pagination';
 import { useFetchMyPlaces } from '@/hooks/useFetchMyPlaces';
-import { useDeletePlace } from '@/hooks/useDeletePlace';
+import { deletePlace } from '@/utils/deletePlace';
 import { useGetMyPlacesCount } from '@/hooks/useGetMyPlacesCount';
 import { AuthContext } from '@/contexts/AuthContext';
 import { PER_PAGE } from '@/constants/perPage';
 
 const Page = () => {
+  const [deleteError, setDeleteError] = useState<string | null>(null);
+  const [isDeleting, setIsDeleting] = useState(false);
   const { user } = useContext(AuthContext);
   const uid = user?.uid;
   const router = useRouter();
-  const { deletePlace, isLoading, deleteError } = useDeletePlace();
 
   const handleDeletePlace = async (id: string) => {
-    await deletePlace(id);
+    if (isDeleting) return;
 
-    if (!error) {
+    setIsDeleting(true);
+    setDeleteError(null);
+
+    try {
+      await deletePlace(id);
       fetchMyPlaces(currentPage, uid || '');
       getCount(uid);
+    } catch (error) {
+      setDeleteError('장소를 삭제하는 중 오류가 발생했습니다.');
+    } finally {
+      setIsDeleting(false);
     }
   };
 
@@ -95,7 +104,7 @@ const Page = () => {
             <MyPlaceCard
               {...place}
               onDelete={handleDeletePlace}
-              isLoading={isLoading}
+              isLoading={isDeleting}
             />
           </div>
         ))}
